Renamed Python 2 generator
