+++ json_codegen/cli.py
@@ -9,6 +9,7 @@ from json_codegen.core import load_schema, load_external_generator
 
 LANGUAGES = {
     "python2": generators.Python2Generator,
+    "python3": generators.Python3Generator,
     "javascript+flow": generators.JavaScriptFlowGenerator,
     "flow": generators.FlowGenerator,
 }
