+++ json_codegen/generators/__init__.py
@@ -1,3 +1,4 @@
-from .python import Python2Generator
+from .python2 import Python2Generator
+from .python3 import Python3Generator
 from .javascript_flow import JavaScriptFlowGenerator
 from .flow import FlowGenerator
