+++ tests/generator_python2_test.py
@@ -4,9 +4,8 @@ from pathlib2 import Path
 import pytest
 import astor
 
-
 from json_codegen import load_schema
-from json_codegen.generators.python import Python2Generator
+from json_codegen.generators.python2 import Python2Generator
 
 
 SCHEMAS_DIR = Path(__file__).parent / "fixtures" / "schemas"
